Track MP4 evaluation video files with Git LFS
Extracts video files formerly in the prior commit (git history which was not
initially made public).  Tracking large binary files slowed down the entire
repository to the point of being unusable with tools such as tig.

Tracks the files using Git Large File Storage [1] [2], which appears to be
preferred to Git Annex as of writing.

[1] https://github.com/git-lfs/git-lfs/wiki/Installation
[2] https://git-lfs.github.com/